int to uint64

+++ src/filter.h
@@ -25,7 +25,7 @@ public:
         filter_[0] = 1;
         for (int i = 0; i < phase_.size(); ++i) {
             if (NonZero(phase_[i])) {
-                std::unordered_map<int, complex_t> updated_filter;
+                std::unordered_map<uint64_t, complex_t> updated_filter;
                 for (auto it : filter_) {
                     int index = it.first;
                     updated_filter[index] = filter_[index] / complex_t(2, 0) + phase_[i] * filter_[(index + (size >> (i + 1))) % size] / complex_t(2, 0);
@@ -37,7 +37,7 @@ public:
         }
     }
 
-    const std::unordered_map<int, complex_t>& FilterTime() const {
+    const std::unordered_map<uint64_t, complex_t>& FilterTime() const {
         return filter_;
     }
 
@@ -55,5 +55,5 @@ private:
 
     int64_t size_;
     std::vector<complex_t> phase_;
-    std::unordered_map<int, complex_t> filter_;
+    std::unordered_map<uint64_t, complex_t> filter_;
 };
